Revert "Host a Service form done, with zod form validation"
This reverts commit 51fdde048cdf07380186c397f7fb9b9f539a6640.
