+++ frontend/starkbnb-web/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.9.1",
     "@starknet-io/types-js": "^0.7.7",
     "@starknet-react/chains": "^3.1.0",
     "@starknet-react/core": "^3.6.0",
@@ -17,11 +16,9 @@
     "next": "15.0.3",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-dom": "19.0.0-rc-66855b96-20241106",
-    "react-hook-form": "^7.53.2",
     "react-icons": "^5.3.0",
     "starknet": "^6.11.0",
-    "starknetkit": "^2.3.3",
-    "zod": "^3.23.8"
+    "starknetkit": "^2.3.3"
   },
   "devDependencies": {
     "@types/node": "^20",
