--- frontend/starkbnb-web/vercel.json
@@ -1,7 +1,0 @@
-{
-    "version": 2,
-    "buildCommand": "npm install --legacy-peer-deps && npm run build",
-    "installCommand": "npm install --legacy-peer-deps",
-    "framework": "nextjs",
-    "outputDirectory": ".next"
-}
