+++ vercel.json
@@ -1,0 +1,16 @@
+{
+    "version": 2,
+    "builds": [
+        {
+            "src": "frontend/starkbnb-web/package.json",
+            "use": "@vercel/static-build"
+        }
+    ],
+    "framework": "nextjs",
+    "routes": [
+        {
+            "src": "/(.*)",
+            "dest": "frontend/starkbnb-web"
+        }
+    ]
+}
